Trigger on use, triggers can only happen once per frame
This prevents multiple triggers associated with the same
character in the level file from triggering at the same time.
For example, if 2 triggers are associated with "a", and both
want to be activated on use key pressed, originally both
would have triggered the function on that frame. Now only
one will trigger the function on the frame the use key is
pressed.